Disable abort test on valgrind
The crash reports cause false-positive warnings when run with valgrind.

+++ tests/integration/logging.tcl
@@ -1,7 +1,7 @@
-set server_path [tmpdir server.log]
 set system_name [string tolower [exec uname -s]]
 
 if {$system_name eq {linux} || $system_name eq {darwin}} {
+    set server_path [tmpdir server.log]
     start_server [list overrides [list dir $server_path]] {
         test "Server is able to generate a stack trace on selected systems" {
             r config set watchdog-period 200
@@ -21,15 +21,20 @@ if {$system_name eq {linux} || $system_name eq {darwin}} {
             }
         }
     }
-}
 
-set server_path [tmpdir server1.log]
-start_server [list overrides [list dir $server_path]] {
-    test "Crash report generated on SIGABRT" {
-        set pid [s process_id]
-        exec kill -SIGABRT $pid
-        set pattern "*STACK TRACE*"
-        set result [exec tail -1000 < [srv 0 stdout]]
-        assert {[string match $pattern $result]}
+    if {!$::valgrind} {
+        set server_path [tmpdir server1.log]
+        start_server [list overrides [list dir $server_path]] {
+            test "Crash report generated on SIGABRT" {
+                r config set crash-memcheck-enabled no
+                r config set use-exit-on-panic yes
+                set pid [s process_id]
+                exec kill -SIGABRT $pid
+                set pattern "*STACK TRACE*"
+                set result [exec tail -1000 < [srv 0 stdout]]
+                assert {[string match $pattern $result]}
+            }
+        }
     }
+
 }
